feat: creating works erros validaiton

+++ src/modules/works/data/services/delete-work.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { DeleteWork } from "../../domain/usecases";
 import { WorkRepository } from "../contracts";
+import { RegisterNotFound } from "src/modules/chore/errors";
 
 @Injectable()
 export class DeleteWorkService implements DeleteWork {
@@ -9,6 +10,12 @@ export class DeleteWorkService implements DeleteWork {
     ) {}
 
     async execute(workId: string): Promise<void> {
+        const workExists = await this.workRepository.getById(workId);
+
+        if(!workExists) {
+            throw new RegisterNotFound('atividade')
+        }
+
         await this.workRepository.delete(workId);
     }
 }
